pyproject.toml
[project]
name = "autoetl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.1",
    "black>=25.9.0",
    "celery>=5.5.3",
    "dask[complete]>=2023.1.0",
    "fastapi>=0.117.1",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "langchain>=1.0.3",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "redis>=7.0.1",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
